Use separate helper class to handle RetroView
