+++ app/src/main/java/com/draco/libretrowrapper/fragments/RetroViewFragment.kt
@@ -160,39 +160,6 @@ class RetroViewFragment : Fragment() {
         }
     }
 
-    fun saveTempState() {
-        /* Save a temporary state since Android killed the activity */
-        val savedInstanceStateBytes = retroView?.serializeState()
-        if (savedInstanceStateBytes != null) {
-            with(privateData.savedInstanceState.outputStream()) {
-                write(savedInstanceStateBytes)
-                close()
-            }
-        }
-    }
-
-    fun restoreTempState() {
-        /* Don't bother restoring a temporary state if it doesn't exist */
-        if (!privateData.savedInstanceState.exists())
-            return
-
-        Thread {
-            /* Wait for the GLRetroView to become usable */
-            retroViewReadyLatch.await()
-
-            /* Fetch the state bytes */
-            val stateInputStream = privateData.savedInstanceState.inputStream()
-            val stateBytes = stateInputStream.readBytes()
-            stateInputStream.close()
-
-            /* Invalidate the temporary state so we cannot restore it twice */
-            privateData.savedInstanceState.delete()
-
-            /* Restore the temporary state */
-            retroView!!.unserializeState(stateBytes)
-        }.start()
-    }
-
     override fun onStop() {
         /* Save emulator settings for next launch */
         saveSettings()
